requirements.txt
python-binance>=1.0.19,<2.0.0
python-dotenv>=1.0.1,<2.0.0
typer>=0.12.3,<1.0.0
rich>=13.7.1,<14.0.0
requests>=2.31.0,<3.0.0